Add Composer to project w/common Composer packages
These commands are run inside the container so the ownership is correct.
The semicolons at the end of lines help Neovim highlight the syntax of
the file correctly.

The consensus about RUN commands is that they should be combined, to
reduce the number of layers created. However, when something goes wrong
the different layers help you to figure out where something went wrong.

NOTE: the package name to install the zip libraries changed at PHP 7.3.

For PHP < 7.3

    RUN apt-get install -y \
            libzip-dev \
            zip \
      && docker-php-ext-configure zip --with-libzip \
      && docker-php-ext-install zip

For PHP >= 7.3

    RUN apt-get install -y \
            libzip-dev \
            zip \
      && docker-php-ext-install zip